Move `Docs` to own file
